CI improvements
* Give JAPARI longer power cycles as it's slow

An timed out run for cifar10_japari_b1:
https://ci.chyen.cc/#/builders/21/builds/4/steps/3/logs/stdio

* Use cached testing data when available
* Remove old log archive before creating a new one
* Consistent naming for log files
* Return difference exit codes for crashes and non-termination

+++ data/download-cifar10.sh
@@ -1,5 +1,6 @@
-#!/bin/sh
+#!/bin/bash
 cd data
+[[ -d cifar-10-batches-py ]] && exit 0
 wget https://www.cs.toronto.edu/~kriz/cifar-10-python.tar.gz
 tar xf cifar-10-python.tar.gz
 rm -v cifar-10-python.tar.gz
